+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 3.0.0-dev.3
 - Add `TreeStateMachine.isStarting`.
+- Add `TreeNodeInfo.getChildren`.
 - Experimental: Add `TreeStateMachine.rootNode`.
 
 ## 3.0.0-dev.2